requirements.txt
Flask==3.0.3
joblib==1.3.2
numpy==1.26.4
scikit-learn==1.4.1.post1
pandas==2.2.1
sqlalchemy==2.0.35
psycopg2-binary
requests
kafka-python